@ Live runtime TerrainDiffusion: AI-generated voxel terrain on local GPU (Phase 1+2)
Brings the diffusion terrain model (xandergos/terrain-diffusion-30m) into the UE5
voxel game, running on the local AMD GPU via NNE + DirectML, generating infinite
procedural terrain as a swappable height source.

Core (pure C++17, clang-harness validated):
- mira::tdiff ports: PortableRng (bit-exact), EdmDpmScheduler, Laplacian,
  WeightWindow, DetailBridge, WorldPipeline (single-tile spine), InfiniteTiler,
  SyntheticMap conditioning. Gates 1-3 green (RNG bit-exact; orchestration parity
  1.68e-6 vs reference). New test_tdiff_* harnesses.
- WorldPipeline.get() gains optional world-origin params so multi-tile noise is
  world-positioned/seamless; getSingleTile (golden path) unchanged.

Engine (MiraThalTerrainAI module):
- FNNEUNetRunner (ORT/DirectML GPU runner), FDiffusionDemService (DEM + detail
  bridge), FDiffusionCoarseProvider (multi-tile blend -> coarse DEM), runner adapter.
- TdiffWorldHook: MiraThal.Tdiff.FillRegion console cmd; installs AI DEM as the
  voxel world height source; sea-anchored vertical mapping + height clamp;
  player land-spawn (searches nearby for dry land, falls back to sea).
- TdiffNewGame: MiraThal.NewGame [seed] new-game flow (random seed if omitted).
- Conditioning from synthetic_map_stats.

Fixes:
- Nanite streaming-pool crash (check(WriteIndex==MaxStreamingPages)): pool sized
  at engine init in DefaultEngine.ini [ConsoleVariables] instead of mutated at
  runtime in VoxelNaniteCrust (init-time-only CVars).

Tooling/docs:
- Rebuild-And-Launch-MiraThal.bat (close->rebuild->launch; the plain launcher
  never rebuilt, loading stale DLLs).
- design/TERRAIN_DIFFUSION_*.md phase plans; gitignore the multi-GB VoxelBake junction.

@

+++ ue5/MiraThal/Source/MiraThalVoxel/Public/Core/ChunkCoords.h
@@ -90,10 +90,16 @@ constexpr Vec3i brick_origin_voxel(const Vec3i& brick) {
 // apron'd 34 for a mesh slab). One formula, reused everywhere, so storage order
 // never disagrees between writer and reader.
 
-constexpr int flatten(int x, int y, int z, int side) {
-    return x + y * side + z * side * side;
+// Returns int64: the index can exceed 2^31 for a large `side` (z*side*side overflows int32
+// once side > ~1290). CHUNK/BRICK/slab(34) are tiny so this never bit, but the helper is reused
+// for arbitrary tile_edge_voxels, so we compute in 64-bit to be overflow-proof. Each term is
+// promoted to int64 BEFORE the multiply, so no intermediate int32 product can wrap.
+constexpr int64_t flatten(int x, int y, int z, int side) {
+    return static_cast<int64_t>(x)
+         + static_cast<int64_t>(y) * side
+         + static_cast<int64_t>(z) * side * side;
 }
-constexpr int flatten(const Vec3i& p, int side) { return flatten(p.x, p.y, p.z, side); }
+constexpr int64_t flatten(const Vec3i& p, int side) { return flatten(p.x, p.y, p.z, side); }
 
 // --- Neighbours ---------------------------------------------------------------
 // The 6 face-adjacent unit steps, in the canonical face order used by the mesher
